requirements.txt
yt-dlp
openai-whisper
ollama
python-dotenv
torch
setuptools-rust